MNT #14 re-enable report command
